fix ambiguious call
open3d> /build/source/cpp/open3d/visualization/rendering/filament/FilamentView.cpp: In member function 'void open3d::visualization::rendering::FilamentView::PreRender()':
open3d> /build/source/cpp/open3d/visualization/rendering/filament/FilamentView.cpp:382:34: error: call of overloaded 'SetParameter(const char [11], const double&)' is ambiguous
open3d>   381 |             FilamentMaterialModifier(selected_material, material_handle)
open3d>       |             ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
open3d>   382 |                     .SetParameter("cameraNear", n)
open3d>       |                     ~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~
open3d> In file included from /build/source/cpp/open3d/visualization/rendering/filament/FilamentView.cpp:37:
open3d> /build/source/cpp/open3d/visualization/rendering/filament/FilamentEntitiesMods.h:40:23: note: candidate: 'virtual open3d::visualization::rendering::MaterialModifier& open3d::visualization::rendering::FilamentMaterialModifier::SetParameter(const char*, int)'
open3d>    40 |     MaterialModifier& SetParameter(const char* parameter, int value) override;
open3d>       |                       ^~~~~~~~~~~~
open3d> /build/source/cpp/open3d/visualization/rendering/filament/FilamentEntitiesMods.h:41:23: note: candidate: 'virtual open3d::visualization::rendering::MaterialModifier& open3d::visualization::rendering::FilamentMaterialModifier::SetParameter(const char*, float)'
open3d>    41 |     MaterialModifier& SetParameter(const char* parameter, float value) override;
open3d>       |                       ^~~~~~~~~~~~

+++ cpp/open3d/visualization/rendering/filament/FilamentView.cpp
@@ -385,8 +385,8 @@ void FilamentView::PreRender() {
             const auto n = camera_->GetNativeCamera()->getNear();
 
             FilamentMaterialModifier(selected_material, material_handle)
-                    .SetParameter("cameraNear", n)
-                    .SetParameter("cameraFar", f)
+                    .SetParameter("cameraNear", static_cast<float>(n))
+                    .SetParameter("cameraFar", static_cast<float>(f))
                     .Finish();
         }
     } else if (mode_ == Mode::Normals) {
